requirements.txt
pandas
PyStemmer